+++ packages/api-models/src/types/Image.ts
@@ -1,4 +1,4 @@
-import type { ISOTimestamp, LicenseURL, RasterMediaType, URL, VectorMediaType } from "@phylopic/utils"
+import type { ISOTimestamp, LicenseURL, RasterMediaType, Tag, URL, VectorMediaType } from "@phylopic/utils"
 import { Entity } from "./Entity"
 import { Link } from "./Link"
 import { Links } from "./Links"
@@ -23,5 +23,6 @@ export interface Image extends Entity<ImageLinks> {
     readonly modified: ISOTimestamp
     readonly modifiedFile: ISOTimestamp
     readonly sponsor: string | null
+    readonly tags: readonly Tag[]
     readonly unlisted?: true
 }